add MIT license to rtoolbox metadata
The `atty.rs` file (copied from the old `atty` crate) is distributed under the terms of the MIT license. This needs to be reflected in crate metadata.

The MIT license also requires that a copy of the license text is included with distributed sources, but I'm not sure if the header in the `src/atty.rs` file is enough to satisfy this requirement.